+++ src/dipdup/context.py
@@ -45,8 +45,10 @@ from dipdup.config.tezos_token_transfers import TezosTokenTransfersIndexConfig
 from dipdup.datasources import Datasource
 from dipdup.datasources import IndexDatasource
 from dipdup.datasources.coinbase import CoinbaseDatasource
+from dipdup.datasources.evm_blockvision import EvmBlockvisionDatasource
 from dipdup.datasources.evm_etherscan import EvmEtherscanDatasource
 from dipdup.datasources.evm_node import EvmNodeDatasource
+from dipdup.datasources.evm_sourcify import EvmSourcifyDatasource
 from dipdup.datasources.evm_subsquid import EvmSubsquidDatasource
 from dipdup.datasources.http import HttpDatasource
 from dipdup.datasources.ipfs import IpfsDatasource
@@ -491,7 +493,7 @@ class DipDupContext:
             defaults={'metadata': metadata, 'update_id': update_id},
         )
 
-    def _get_datasource(self, name: str, *types: type[DatasourceT]) -> DatasourceT:
+    def get_datasource(self, name: str, *types: type[DatasourceT]) -> DatasourceT:
         datasource = self.datasources.get(name)
         if not datasource:
             raise ConfigurationError(f'Datasource `{name}` is missing')
@@ -506,67 +508,75 @@ class DipDupContext:
 
     def get_tezos_tzkt_datasource(self, name: str) -> TezosTzktDatasource:
         """Get `tezos.tzkt` datasource by name"""
-        return self._get_datasource(name, TezosTzktDatasource)
+        return self.get_datasource(name, TezosTzktDatasource)
 
     def get_evm_subsquid_datasource(self, name: str) -> EvmSubsquidDatasource:
         """Get `evm.subsquid` datasource by name"""
-        return self._get_datasource(name, EvmSubsquidDatasource)
+        return self.get_datasource(name, EvmSubsquidDatasource)
 
     def get_evm_node_datasource(self, name: str) -> EvmNodeDatasource:
         """Get `evm.node` datasource by name"""
-        return self._get_datasource(name, EvmNodeDatasource)
+        return self.get_datasource(name, EvmNodeDatasource)
 
     def get_evm_etherscan_datasource(self, name: str) -> EvmEtherscanDatasource:
         """Get `evm.etherscan` datasource by name
 
         :param name: Name of the datasource
         """
-        return self._get_datasource(name, EvmEtherscanDatasource)
+        return self.get_datasource(name, EvmEtherscanDatasource)
 
     # NOTE: Alias, remove in 9.0
     get_abi_etherscan_datasource = get_evm_etherscan_datasource
 
-    def get_evm_datasource(self, name: str) -> EvmSubsquidDatasource | EvmNodeDatasource | EvmEtherscanDatasource:
+    def get_evm_datasource(
+        self, name: str
+    ) -> (
+        EvmSubsquidDatasource
+        | EvmNodeDatasource
+        | EvmEtherscanDatasource
+        | EvmSourcifyDatasource
+        | EvmBlockvisionDatasource
+    ):
         """Get `evm` datasource by name"""
-        return self._get_datasource(name, EvmSubsquidDatasource, EvmNodeDatasource, EvmEtherscanDatasource)  # type: ignore[return-value]
+        return self.get_datasource(name, EvmSubsquidDatasource, EvmNodeDatasource, EvmEtherscanDatasource, EvmSourcifyDatasource, EvmBlockvisionDatasource)  # type: ignore[return-value]
 
     def get_starknet_datasource(self, name: str) -> StarknetSubsquidDatasource | StarknetNodeDatasource:
         """Get `starknet` datasource by name"""
-        return self._get_datasource(name, StarknetSubsquidDatasource, StarknetNodeDatasource)  # type: ignore[return-value]
+        return self.get_datasource(name, StarknetSubsquidDatasource, StarknetNodeDatasource)  # type: ignore[return-value]
 
     def get_substrate_datasource(
         self, name: str
     ) -> SubstrateSubsquidDatasource | SubstrateSubscanDatasource | SubstrateNodeDatasource:
         """Get `substrate` datasource by name"""
-        return self._get_datasource(name, SubstrateSubsquidDatasource, SubstrateSubscanDatasource, SubstrateNodeDatasource)  # type: ignore[return-value]
+        return self.get_datasource(name, SubstrateSubsquidDatasource, SubstrateSubscanDatasource, SubstrateNodeDatasource)  # type: ignore[return-value]
 
     def get_coinbase_datasource(self, name: str) -> CoinbaseDatasource:
         """Get `coinbase` datasource by name
 
         :param name: Name of the datasource
         """
-        return self._get_datasource(name, CoinbaseDatasource)
+        return self.get_datasource(name, CoinbaseDatasource)
 
     def get_tzip_metadata_datasource(self, name: str) -> TzipMetadataDatasource:
         """Get `metadata` datasource by name
 
         :param name: Name of the datasource
         """
-        return self._get_datasource(name, TzipMetadataDatasource)
+        return self.get_datasource(name, TzipMetadataDatasource)
 
     def get_ipfs_datasource(self, name: str) -> IpfsDatasource:
         """Get `ipfs` datasource by name
 
         :param name: Name of the datasource
         """
-        return self._get_datasource(name, IpfsDatasource)
+        return self.get_datasource(name, IpfsDatasource)
 
     def get_http_datasource(self, name: str) -> HttpDatasource:
         """Get `http` datasource by name
 
         :param name: Name of the datasource
         """
-        return self._get_datasource(name, HttpDatasource)
+        return self.get_datasource(name, HttpDatasource)
 
     async def rollback(self, index: str, from_level: int, to_level: int) -> None:
         """Rollback index to a given level reverting all changes made since that level.